Make teacher portal a separate page, move the home logic out to default layout
