+++ components/TeacherPortal/teacher-portal.state.tsx
@@ -1,0 +1,3 @@
+export const portalState = {
+  showing: 'one-time-code' as 'one-time-code' | 'shared-message',
+};
